fixeo prototipos paginacion

+++ mi-ram-hq/include/mi-ram-hq-lib.h
@@ -86,7 +86,7 @@ typedef struct{
 
 typedef struct{
 	t_pagina* pagina;
-	double indice;
+	int indice;
 	int offset;
 }inicio_tcb;
